+++ enum-table/src/impls/mod.rs
@@ -1,0 +1,4 @@
+mod std;
+
+#[cfg(feature = "serde")]
+mod serde;
